requirements.txt
psycopg2==2.9.5
pyodbc==4.0.35
python-dotenv==0.21.0
